+++ src/main/java/dan200/computercraft/shared/computer/core/ServerComputer.java
@@ -290,7 +290,7 @@ public class ServerComputer extends ServerTerminal
     @Override
     public double getTimeOfDay()
     {
-        return (double)((m_world.getWorldTime() + 6000) % 24000) / 1000.0;
+        return (m_world.getWorldTime() + 6000) % 24000 / 1000.0;
     }
 
     @Override
